[RFC] Approach 3

+++ org.eclipse.jgit/src/org/eclipse/jgit/internal/storage/file/RefDirectory.java
@@ -712,7 +712,7 @@ public class RefDirectory extends RefDatabase {
 		int levels = levelsIn(name) - 2;
 		delete(logFor(name), levels);
 		if (dst.getStorage().isLoose()) {
-			deleteAndUnlock(fileFor(name), levels, update);
+			deleteAndUnlock(fileFor(name), update);
 		}
 
 		modCnt.incrementAndGet();
@@ -1116,7 +1116,12 @@ public class RefDirectory extends RefDatabase {
 			if (n == null) {
 				if (looseRefs.compareAndSet(curList, curList.remove(idx)))
 					modCnt.incrementAndGet();
-				return packed.get(name);
+				Ref packedRef = packed.get(name);
+				if (packedRef != null && hasTombstone(fileFor(name))) {
+					return null;
+				} else {
+					return packedRef;
+				}
 			}
 
 			if (o == n)
@@ -1128,13 +1133,13 @@ public class RefDirectory extends RefDatabase {
 
 		final LooseRef n = scanRef(null, name);
 		if (n == null)
-			return packed.get(name);
+			return hasTombstone(fileFor(name)) ? null : packed.get(name);
 
 		// check whether the found new ref is the an additional ref. These refs
 		// should not go into looseRefs
 		for (String additionalRefsName : additionalRefsNames) {
 			if (name.equals(additionalRefsName)) {
-				return n;
+				return hasTombstone(fileFor(name)) ? null : n;
 			}
 		}
 
@@ -1352,18 +1357,42 @@ public class RefDirectory extends RefDatabase {
 		}
 	}
 
-	private static void deleteAndUnlock(File file, int depth,
-			RefDirectoryUpdate refUpdate) throws IOException {
+	private static void deleteAndUnlock(File file, RefDirectoryUpdate refUpdate) throws IOException {
+		if (!createTombstone(file)) {
+			throw new IOException("Unable to create tombstone for file " + file.getAbsolutePath());
+		}
 		delete(file);
 		if (refUpdate != null) {
 			refUpdate.unlock(); // otherwise cannot delete parent directories emptied by the update
 		}
-		deleteEmptyParentDirs(file, depth);
+	}
+
+	private static boolean createTombstone(File file) throws IOException {
+		return getTombstoneFile(file).createNewFile();
+	}
+
+	private static File getTombstoneFile(File file) {
+		return new File(file.getParent() + "/." + file.getName());
+	}
+
+	private static boolean deleteTombstone(File file, int depth) {
+		boolean deleted = getTombstoneFile(file).delete();
+		if (deleted) {
+			deleteEmptyParentDirs(file, depth);
+		}
+		return deleted;
+	}
+
+	private static boolean hasTombstone(File file) {
+		return getTombstoneFile(file).exists();
 	}
 
 	private static void deleteAndUnlock(File file, int depth, LockFile rLck)
 			throws IOException {
 		delete(file);
+		if (!deleteTombstone(file, depth)) {
+			throw new IOException("Unable to remove ref tombstone " + file.getAbsolutePath());
+		}
 		if (rLck != null) {
 			rLck.unlock(); // otherwise cannot delete parent directories of the lock file
 		}
